Moving overlay component to ds
