+++ src/components/ds/overlay/overlay.tsx
@@ -1,8 +1,8 @@
 import classNames from "classnames";
-import useNavigationStore from "../../hooks/navigation-store";
-import IconButton from "../ds/button/icon-button";
-import CloseIcon from "../ds/icons/close-icon";
-import { Loading } from "../ds/icons/loading";
+import useNavigationStore from "../../../hooks/navigation-store";
+import IconButton from "../button/icon-button";
+import CloseIcon from "../icons/close-icon";
+import { Loading } from "../icons/loading";
 import styles from "./overlay.module.css";
 
 const Overlay: React.FC<{
